refactor(express-example): separate app creation from server execution
Remove awkward import.meta.url conditional by splitting into two files:
- src/app.ts: Pure app creation (no side effects)
- src/server.ts: Executable entry point (starts server)

Benefits:
- No import.meta.url hack needed
- Single setup path for tests and server
- Clear separation of concerns
- Standard Node.js pattern

Tests import from app.ts, server.ts is executable only.
All 85 tests passing.
