+++ apps/express-example/src/server.ts
@@ -1,87 +1,12 @@
-import {
-  createScenarist,
-  type ExpressScenarist,
-} from "@scenarist/express-adapter";
-import express, { type Express } from "express";
-import { setupGitHubRoutes } from "./routes/github.js";
-import { setupStripeRoutes } from "./routes/stripe.js";
-import { setupWeatherRoutes } from "./routes/weather.js";
-import { setupCartRoutes } from "./routes/cart.js";
-import { setupFormRoutes } from "./routes/form.js";
-import { setupStringMatchingRoutes } from "./routes/string-matching.js";
-import { setupUrlMatchingRoutes } from "./routes/url-matching.js";
-import { setupHostnameMatchingRoutes } from "./routes/hostname-matching.js";
-import { setupProductsRepoRoutes } from "./routes/products-repo.js";
-import { scenarios } from "./scenarios.js";
+import { createApp } from "./app.js";
 
 /**
- * Create and configure the Express application with Scenarist
+ * Start the Express server with Scenarist
  *
- * **PRODUCTION TREE-SHAKING (Automatic):**
- *
- * Scenarist adapters automatically optimize for production:
- * - In production: `createScenarist()` returns `undefined`
- * - Bundlers detect this and eliminate ~300KB of test code
- * - No configuration needed - it just works!
- *
- * **How it works:**
- * 1. Adapter checks `process.env.NODE_ENV === 'production'`
- * 2. Returns `undefined` without loading any Scenarist code
- * 3. Bundlers perform dead code elimination
- * 4. Result: Zero test code in production bundles
- *
- * **Type safety:**
- * - Return type: `ExpressScenarist<typeof scenarios> | undefined`
- * - TypeScript enforces null checks via `if (scenarist)` guards
- *
- * **Learn more:** https://scenarist.io/introduction/production-safety
+ * This file is the entry point for running the server directly.
+ * Tests should import from './app.js' instead.
  */
-export const createApp = async (): Promise<{
-  app: Express;
-  scenarist: ExpressScenarist<typeof scenarios> | undefined;
-}> => {
-  const app = express();
-
-  // Parse JSON bodies
-  app.use(express.json());
-
-  // Initialize Scenarist (automatically returns undefined in production)
-  const scenarist = await createScenarist({
-    enabled: true,
-    scenarios, // All scenarios registered at initialization (must include 'default')
-    strictMode: false, // Allow passthrough for unmocked requests
-  });
-
-  // Apply Scenarist middleware (only in non-production)
-  if (scenarist) {
-    app.use(scenarist.middleware);
-  }
-
-  // Setup application routes
-  const router = express.Router();
-  setupGitHubRoutes(router);
-  setupWeatherRoutes(router);
-  setupStripeRoutes(router);
-  setupCartRoutes(router);
-  setupFormRoutes(router);
-  setupStringMatchingRoutes(router);
-  setupUrlMatchingRoutes(router);
-  setupHostnameMatchingRoutes(router);
-  setupProductsRepoRoutes(router);
-  app.use(router);
-
-  // Health check endpoint
-  app.get("/health", (_req, res) => {
-    res.json({ status: "ok" });
-  });
-
-  return { app, scenarist };
-};
-
-/**
- * Start the server (only if running directly, not when imported for tests)
- */
-if (import.meta.url === `file://${process.argv[1]}`) {
+const main = async () => {
   const { app, scenarist } = await createApp();
 
   // Start MSW (only in non-production)
@@ -96,26 +21,6 @@ if (import.meta.url === `file://${process.argv[1]}`) {
       console.log(`Scenario control: POST http://localhost:${PORT}/__scenario__`);
     }
   });
-}
-/**
- * Start the server (only if running directly, not when imported for tests)
- */
-if (import.meta.url === `file://${process.argv[1]}`) {
-  const main = async () => {
-    const { app, scenarist } = await createApp();
+};
 
-    // Start MSW (only in non-production)
-    if (scenarist) {
-      scenarist.start();
-    }
-
-    const PORT = process.env.PORT || 3000;
-    app.listen(PORT, () => {
-      console.log(`Server running on http://localhost:${PORT}`);
-      if (scenarist) {
-        console.log(`Scenario control: POST http://localhost:${PORT}/__scenario__`);
-      }
-    });
-  };
-  main().catch(console.error);
-}
+main().catch(console.error);
